+++ sapiens.sh
@@ -20,117 +20,22 @@ case $1 in
         if [ -f "$CONFIG_FILE" ]; then
             source "$CONFIG_FILE"
         else
-            logit "ERROR" "$CONFIG_FILE not found. Please run ./sapiens.sh install to configure an active world."
-            exit 1
+            logit "ERROR" "$CONFIG_FILE not found. Please run ./sapiens.sh install to configure an active world." "echo"
+            read -p "$(echo -e "${YELLOW}Do you want to run the installation now? (${GREEN}y${YELLOW}/${RED}n${YELLOW}): ${NC}")" choice
+            case "$choice" in 
+                y|Y ) install_server;;
+                n|N|* ) echo -e "${RED}Installation aborted. Exiting...${NC}"; exit 1;;
+            esac
         fi
         ;;
 esac
 
 case $1 in
     install)
-        logit "DEBUG" "*********************** install.sh started ***********************"
-        logit "DEBUG" "Calling check_for_root"; check_for_root
-
-        # Show the welcome screen
-        splash_text
-
-        # Check to see if the software dependencies are in place
-        if ! dependencies_installed; then
-            logit "INFO" "Dependency check failed on linux account $(whoami)."
-            echo "The account $(whoami) does not have the necessary software installed to run the Sapiens linuxServer."
-            if ! yesno "Would you like to install it now?"; then
-                echo "Installation aborted."
-                logit "INFO" "Dependency installation aborted by user."
-                exit 0
-            else
-                logit "DEBUG" "Calling install_dependencies"
-                install_dependencies    # Install Steamcmd and the other required dependencies
-                logit "DEBUG" "Calling patch_steam"
-                patch_steam             # Patch for the steam client.
-                logit "DEBUG" "Calling upgrade_sapiens"
-                upgrade_sapiens         # Use steamcmd to update the sapiens executable.
-                echo "Sapiens Server Manager installation successfully complete!"
-                read -n 1 -s -r -p "Press any key to continue"
-                echo ""  # Move to the next line after the key press
-            fi
-        fi
-
-        # Check to see if the linuxServer executable is present
-        if ! sapiens_installed; then
-            logit "INFO" "Sapiens linuxServer not found.  Installing."
-            echo "Sapiens linuxServer not found.  Installing."
-            upgrade_sapiens
-        fi
-
-        # Ensure the needed directories exist
-        set_permissions
-        add_to_path $SCRIPT_DIR # Add the script directory to the path.
-
-        while true; do
-            # Call installed_worlds and store the count
-            world_count=$(installed_worlds)
-
-            echo -e "${BRIGHT_CYAN}-------------------------------------------------------${NC}"
-            # Check if the world count is greater than zero
-            if [[ $world_count -gt 0 ]]; then
-                echo -e "${BRIGHT_GREEN}***** $world_count detected installed worlds *****${NC}\n"
-                show_installed_worlds "clean"
-            else
-                echo -e "${RED}No installed worlds detected.${NC}"
-            fi
-            echo -e "${BRIGHT_CYAN}-------------------------------------------------------${NC}"
-            echo -e "${BRIGHT_YELLOW}1. Make an existing world active${NC}"
-            echo -e "${BRIGHT_YELLOW}2. Create a new world${NC}"
-            echo -e "${BRIGHT_YELLOW}0. Exit${NC}"
-            echo -e "${BRIGHT_CYAN}-------------------------------------------------------${NC}"
-            read -p "Enter your choice (1 existing, 2 for new, 0 to Exit): " user_choice
-
-            case $user_choice in
-                1)
-                    select_world
-                    if [ $? -eq 0 ]; then
-                        echo "World selected : $WORLD_NAME"
-                        echo "World ID       : $WORLD_ID"
-                        echo "Server ID      : $SERVER_ID"
-                        logit "INFO" "World selected: $WORLD_NAME"
-                        logit "INFO" "World ID      : $WORLD_ID"
-                        logit "INFO" "Server ID     : $SERVER_ID"
-                        # There was an existing world that will be used.
-                    else
-                        echo "Failed to select a world."
-                        logit "DEBUG" "User made an invalid select_world menu choice."
-                        exit 1
-                    fi
-                    ;;
-                2)
-                    # Calls function to create a new world
-                    WORLD_ID=$(create_world)
-                    if [[ $WORLD_ID ]]; then
-                        logit "INFO" "World created and selected: $WORLD_NAME (ID: $WORLD_ID)"
-                    else
-                        logit "ERROR" "Failed to create a new world."
-                        continue # Skip the menu refresh if world creation failed
-                    fi
-                    ;;
-                0)
-                    echo "Exiting program."
-                    exit 0
-                    ;;
-                *)
-                    echo "Invalid choice. Please enter 1 or 2."
-                    ;;
-            esac
-
-            # Break loop if not creating a new world to avoid showing the menu again unnecessarily
-            if [[ "$user_choice" != "2" ]]; then
-                break
-            fi
-        done
-
-        get_active_server_details
-        create_config
+        install_server
         ;;
     start)
+        upgrade_server          # Check to see if we have the latest version.
         start_server            # manage_world.sh
         ;;
     stop)
@@ -142,7 +47,6 @@ case $1 in
     restart)
         logit "INFO" "Restarting server"
         stop_server             # manage_world.sh
-        sleep 5  # wait for the server to shut down gracefully
         start_server            # manage_world.sh
         ;;
     autorestart)
@@ -156,23 +60,27 @@ case $1 in
         ;;
     console)
         open_console
-        ;;    
+        ;;
     *)
+        clear
         echo "$SCRIPT_NAME"
         echo "Script Version: $SCRIPT_VERSION"
         echo "Author: chillgenxer@chillgenxer.com"
         echo "GitHub: $GITHUB_URL"
         echo ""
+        echo "Run './sapiens.sh install' to install dependencies and the Sapiens Server."
+        echo ""
         echo "Usage examples:"
+        echo "---------------"
         echo "./sapiens.sh start - starts your world in the background."
         echo "./sapiens.sh console - Bring the running world's console. To exit without stopping the server hold CTRL and type A D."        
-        echo "./sapiens.sh stop - stops your world."
+        echo "./sapiens.sh stop - Stops the world."
+        echo "./sapiens.sh hardstop - Stops the world and cancels any autorestart setting."
         echo "./sapiens.sh restart - Manually restart the server. Good to use if things are getting laggy."
-        echo "./sapiens.sh autorestart [hours] - Automatically restart the world at the specified hour interval."
-        echo "./sapiens.sh upgrade - This will update you to the latest version of the Sapiens server."
+        echo "./sapiens.sh autorestart [0-24] - Automatically restart the world at the specified hour interval, 0 cancels."
+        echo "./sapiens.sh upgrade - Upgrade to the latest version of the Sapiens server executable from Steam."
         echo "./sapiens.sh backup - Stops the world and backs it up to the backup folder."
-        echo "./sapiens.sh install - Installs dependencies and configures the Sapiens server."
-
+        echo ""
         exit 1
         ;;
 esac