+++ changelog.md
@@ -1,8 +1,7 @@
 ##1.0.0##
 - Updated enet log location and added sapservermgr.log
 - Refactored code
-- 
-
+- Checks Sapiens version on startup and upgrades if necessary.
 
 ##0.4.1## - 2024-05-05
 - Made bug reporting options configurable.